pyproject.toml
[tool.poetry]
name = "onlayn-kurs-sayt"
version = "0.1.0"
description = ""
authors = ["shahzod <qoshqulovshahzod9@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.6"
django-mptt = "^0.16.0"
djangorestframework-simplejwt = "^5.4.0"
django-ckeditor-5 = "^0.2.17"
django-rest-swagger = "^2.2.0"
drf-spectacular = "^0.28.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
